Update student_login.html and passes/views.py
Modified the student_login page to render all selected destinations by the user, and for each destination will display all passes associated with it, as well as a button to send passes to that destination.

Also modified the views to redirect to the monitoring page after selecting destinations. As well as hooking up the generated buttons per destination to send post requests, and upload the correct hall pass data